Fix back navigation issue on signup screen
When user is in the EmailPasswordsScreen.kt, the user has the option to go back to login screen. Whenever user clicks that "Login" text and decides to use back navigation, it will go back to the signup screen, which should not be the intended behavior. This commit fixes this by ensuring that the screen won't go back to the EmailPasswordsScreen.kt when user uses back navigation.